Move shared (firestore) types to functions and reference from top level app (as single source of truth)
