+++ eisbuk-admin/src/components/slots/SlotListByDay/Slot.tsx
@@ -20,9 +20,9 @@ import {
   CheckCircle as CheckCircleIcon,
 } from "@material-ui/icons";
 
-import { Duration } from "@/enums/firestore";
+import { Duration } from "@functions/enums/firestore";
 
-import { Slot as SlotInterface, BookingInfo } from "@/types/firestore";
+import { Slot as SlotInterface, BookingInfo } from "@functions/types/firestore";
 import { LocalStore } from "@/types/store";
 import { SlotOperation } from "@/types/slotOperations";
