+++ eisbuk-admin/src/components/slots/SlotListByDay/SlotsDay.tsx
@@ -17,8 +17,11 @@ import FileCopyIcon from "@material-ui/icons/FileCopy";
 import AddCircleOutlineIcon from "@material-ui/icons/AddCircleOutline";
 import AssignmentIcon from "@material-ui/icons/Assignment";
 
+import { __isStorybook__ } from "@/lib/constants";
+
 import { LocalStore } from "@/types/store";
-import { Slot as SlotInterface } from "@/types/firestore";
+import { Slot as SlotInterface } from "@functions/types/firestore";
+
 import { ETheme } from "@/themes";
 
 import {
@@ -33,7 +36,6 @@ import CustomerAreaBookingCard from "@/components/customerArea/CustomerAreaBooki
 import Slot, { SlotProps } from "./Slot";
 
 import { shiftSlotsDay } from "@/data/slotutils";
-import { __isStorybook__ } from "@/lib/constants";
 
 const luxon = new LuxonUtils({ locale: "C" });
