+++ eisbuk-admin/src/pages/CustomerAreaPage/CustomerAreaPage.tsx
@@ -10,13 +10,15 @@ import {
 } from "@material-ui/icons";
 import { useTranslation } from "react-i18next";
 
+import { LocalStore } from "@/types/store";
+
+import { OrgSubCollection } from "@functions/enums/firestore";
+
 import AppbarCustomer from "@/components/layout/AppbarCustomer";
 import AppbarAdmin from "@/components/layout/AppbarAdmin";
 import CustomerAreaCalendar from "./CustomerAreaCalendar";
 
 import { wrapOrganization } from "@/utils/firestore";
-import { LocalStore } from "@/types/store";
-import { OrgSubCollection } from "@/enums/firestore";
 
 // ***** Region Link Tab ***** //
 type TabProps = Parameters<typeof Tab>[0];
