+++ eisbuk-admin/src/types/slotOperations.ts
@@ -1,8 +1,8 @@
 import { Timestamp } from "@google-cloud/firestore";
 
-import { SlotType, Category, Duration } from "@/enums/firestore";
+import { SlotType, Category, Duration } from "@functions/enums/firestore";
 
-import { Slot } from "@/types/firestore";
+import { Slot } from "@functions/types/firestore";
 
 /**
  * Params shared between different types of operation on slot ("create" | "edit")
