+++ eisbuk-admin/tsconfig.paths.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"]
+      "@/*": ["src/*"],
+      "@functions/*": ["functions/src/*"]
     }
   }
 }
